+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Ch-ch-ch-ch-changes
 
+## 1.1.5
+
+- Add `--cwd` flag to Launcher to use current directory as working directory.
+
 ## 1.1.4
 
 - Add `--absolute` flag to `nx_tools list`. The goal is to aid in setting `UGII_TMG_DIR` without opening Windows Explorer and navigating through directories. Maybe in the future an `nx_tools set_tmg` command will be added (if requested).